Fixed an issue where the stream/byte[] objects were being GC'd (#3)
The libzip seems to write most of its data during the close
method call. If a stream or byte[] array has been collected
it can cause problems.

Also the callback seems to be called AFTER the ZipArchive has
been disposed. Which causes a SIGSEGV. So we need to make the
callback static.