Change the way ZipArchive.AddFiles works
We used to store files ignoring their directory part which may cause
naming conflicts if there are files with the same name but in different
directories. However, since the behavior of ignoring file directories
may be desirable sometimes, this patch makes the behavior optional.

The default action is to use the source file directory part as a
subdirectory relative to the one passed to AddFiles as the root dir (
it defaults to / if null or empty).

Also, improve documentation.