Properly convert to/from UTF8 in libzip calls
libzip encodes all the strings it returns in UTF-8 and it expects
them to be passed to it in UTF-8 as well. This works (by chance,
to be honest) on Unix since most Unix hosts use UTF-8 right now.
It, however, fails on Windows which uses either Unicode or ANSI.

This patch, based on one submitted by Jérémie Laval in

https://bugzilla.xamarin.com/show_bug.cgi?id=44038#c1

fixes this issue by properly re-encoding the strings to/from the
current system encoding to/from UTF-8.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=44268