Switch the Addfile method over to using the Managed Callback. (#20)
Context https://bugzilla.xamarin.com/show_bug.cgi?id=44184

Currently when adding a File the system was using
libzip's interal callback to add the file. This seems
to make our support a bit harder since we only get
system level error codes if there is an error or
sharing violation.

By switching over to using the Managed callback and
FileStream we should be able to get decent .net
exceptions if we are unable to read a file.