Modify zip-slip fix to preserve actual entry names
The zip-slip fix simply replaced all the \ characters with /, assuming that all
entries in the zip file will follow the ZIP standard and have the correct path
segment separators. Alas, aapt2 on Windows creates entries with mixed / and \
characters and those entries became inaccessible because of the zip-slip fix.

Preserve the actual name but add a new property, `ZipEntry.NativeFullName`,
which contains the OS-native path of the entry while `ZipEntry.FullName`
contains the actual entry from the archive.

`ZipEntry.Extract` got a new parameter, `useNativeFileName` (defaults to
`false`) which tells it to use the OS-native file name on extraction, if set to
`true`.

Add a `ZipEntry.Rename` method to rename an entry easily.