Upgrade to SDK Style project. Add libzip building and nuget
This PR is pulling in the libzip native library, building
both Mac and Windows versions and creating a nuget package.

We are also creating a nuget package for libZipSharp. This
package will depend on the libzip one so that the native
parts will be pulled in automatically in addition to users
being able to use libzip native as a stand alone nuget.

We also added build automation. We support both azure pipelines
and travis ci. The main automation will be done via azure
pipelines.

The C# projects were upgraded to SDK style projects, this was
to aide the generation of the nuget packages.