[windows] remove NuGets for libzip and zlib
The NuGets we were using on Windows:

* Are quite old
* Appear to not work for 32-bit processes

xamarin/xamarin-android builds its own copy of these native libraries,
but in some cases `libZipSharp.csproj` was overwriting them.

We think it's better to just remove these NuGet references for Windows
here, since we are not supplying the native libraries for the other
platforms.