Bump libzip to 1.6.1 (#49)
In order to build libzip 1.6.1 we had to switch
over to using an actual Windows machine. This
was because libzip now no longer compiles in
mingw.

To help with this we are using the excellent
[vcpkg](https://github.com/Microsoft/vcpkg) project.
This project helps us grab and install dependencies
in a cross platform manner. However it is only used
on windows at the moment.