Fix Native Crash on Windows.
We had a problem in Xamarin.Android where libzip
would crash on our test machine. This turned out
to be a bung in libzip itself.

	https://github.com/nih-at/libzip/pull/202

There was a missing `NULL` check in `zip_stat_index`.
This PR adds a patch to allow us to patch this problem
in the latest release of `libzip` so that we can
bump to the latest version in Xamrain.Android.