Fix Windows to look in Assembly Directory for 32bit dll.
Commit 96eb5e34 introduced the use of `DefaultDllImportSearchPathsAttribute`
to add extra security when loading native libraries. This
works fine when the executable that is running is in the
same directory as the `LibZipSharp.dll` and the `libzip.dll`.

However when the executable has a different `WorkingDirectory`
as the `LibZipSharp.dll` this then causes the following
error.

	System.DllNotFoundException: Unable to load DLL 'libzip'

This is because with the new attribute only Safe paths are searched.
So when we run this under MSBuild.exe for example, the `Working`
directory will definately NOT contain the 32bit `libzip.dll`.

The fix to to always set the location of the `libzip.dll` via
`SetDllDirectory` for both 32bit and 64bit processes.

Also bump to 1.0.18