Build on Windows again (#71)
Commit 755a42a8 reverted the system which built the native libzip.dll in Windows.
This was because we suspected that the change from using mingw was causing a runtime
crash. However this turned out not to be the case (see https://github.com/xamarin/LibZipSharp/pull/67).

So this PR adds back in the code which builds the windows dll on windows.

It also updates the libzip-changes.patch so that it would apply on windows via `git apply`.
This patch now works on all platforms.

Also testing out a new release package system. Rather than just publishing to stable on nuget,
we will first publish an `x.x.x-rc` release candidate. This will allow us to do a test PR on
xamarin.android without breaking the current stable package. We had a situation over
the last few months where the "stable" package actually had a crash on windows.
This is not ideal for people consuming the nuget.

So from now on we will publish an `-rc` do a bump in xamarin.android and if that
passes, then we can publish a stable release.