Do not use the XZ library (LZMA) by default
Disable adding support for XZ (LZMA) compression by default. This is due
to the fact that XZ is licensed under `LGPL` and it may adversely affect
the licensing of the shared library we place in the nuget.

XZ can be enabled by passing the `--xz` option to the `build.sh` script
and by passing the `/p:UseXZ=True` option to the managed library build.