[nuget] Include Mono.Unix build assets (#91)
This is to make projects which reference only LibZipSharp work with the
new Mono.Unix package.  By default, the nuget generated for LibZipSharp
will not include Mono.Unix's `build` components which will break said
projects.  Transitive build assets must be included so that the
Mono.Unix configuration and runtime files are copied to the project
directory.