Add a new more complex unit test (#105)
Add a new more complete unit test which emulates how Xamarin.Android uses LibZipSharp.

Also made a small change to the way we handle reading from a stream. According to the documentation stream.Read can return a value less than what was asked. So we need to keep reading until we get all the data we expected or an error is thrown.