Use RuntimeInformation to determine the Processor architecture.
See https://docs.microsoft.com/en-us/dotnet/api/system.runtime.interopservices.architecture?view=net-6.0

Now that we ship an arm based native library on windows we need to be
able to determine the underlying library to load. We can use the
`RuntimeInformation.ProcessArchitecture` value to get this data.

Then based on that we can load from `x86`, `x64` or `arm64` subdirectories.
Note we will still fallback to loading from the current directory , but
the preference is to load from on of the architecture specific sub directories.