Remove LZMA (XZ) support (#139)
Context: https://tukaani.org/xz-backdoor/

In light of the recently discovered backdoor in xz-utils and its
GitHub repository being blocked, remove the submodule from LibZipSharp
and remove all the code to enable lzma compression support in it.

Note that XZ support was disabled by default and was never released
as part of any official LibZipSharp nugets/binaries, therefore we
are not affected in any form or shape by the issue.

The submodule is removed so that it is possible to clone and initialize
this repository.  When the issues surrounding xz-utils are fixed, we can
restore support by reverting this commit.