Fix the Illusive invalid zip archive issue that has been a problem for ages!
Fixes xamarin/xamarin-android#8988

We had this odd corrupt zip file issue which kept cropping up on our
Azure Pipelines builds. We had no idea what caused it until now.
It turns out some of the logic we were using to control the stream
position was not working correctly. It stopped us from navigating to a
0 offset. As a result some of the data for the local headers of an item
(not the central directory) would be written incorrectly.
This would result in a zip which may or may not be extractable, it would
depend on how resilient the software extracting the data would be.

We discovered that just enabling the strict consistency checks would
uncover the issue, so that has been enabled in a number of unit tests.
Once we did that it turns out we were writting the corrupt data
ALL the TIME!. Fixing up the seeking code to take into account that we
might want to see to 0 fixed the issue.