Dev/dellis1972/optimze (#148)
* Make sure we Optimize the assembly

We are seeing the following issue when using Benchmark.net 

```
Assembly ZipBench which defines benchmarks references non-optimized libZipSharp
	If you own this dependency, please, build it in RELEASE.
	If you don't, you can disable this policy by using 'config.WithOptions(ConfigOptions.DisableOptimizationsValidator)
```

which suggests we are not shipping an Optimized assembly. Which is weird...
So we need to turn on Optimize..

Looking at the Benchmark code https://github.com/dotnet/BenchmarkDotNet/blob/master/src/BenchmarkDotNet/Extensions/AssemblyExtensions.cs#L10. We see that the `IsJitOptimizerDisabled` is NOT true in our case. So we have no optimised the assembly 

* v3.5.1

* Bump libzip to v1.11.2

* Bump zlib-ng to 2.2.3