[ci] Add a schedule trigger for OneLocBuild (#151)
Adds a schedule trigger to ensure the OneLocBuild task runs regularly,
which should ensure it stays active to the loc system.

The compliance stage condition has been adjusted so that it will be
skipped when running a scheduled build.

The build will also use the new internal yaml-templates repo.

Other changes:

* Exclude `internal-templates` from source analysis

* Exclude implicit `yaml-templates` repo resources as well